Implemented suggestions from review
